pyproject.toml
################################################################################
# Project
[project]
name = "ros_to_markdown"
version = "0.1.0"
description = "Convert ROS systems to Markdown for LLM processing"
authors = [{ name = "Rob Royce", email = "rob.royce@jpl.nasa.gov" }]
requires-python = ">=3.8,<3.13"
dependencies = [
    "pyyaml",
    "structlog[dev]",
    "colorama>=0.4.0",
    "pydantic",
    "jinja2",
    "clang==13.0.1",
    "networkx",
    "rich",
    "click",
]
readme = "README.md"
license = { text = "Apache 2.0" }
################################################################################


################################################################################
# Development
[project.optional-dependencies]

dev = [
    "black>=23.0",
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.0.290",
    "pytest-mock>=3.10.0",
]
################################################################################


################################################################################
# Build System
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["src/ros_to_markdown"]
################################################################################


################################################################################
# Linting
[tool.ruff]
line-length = 100
target-version = "py38"
################################################################################


################################################################################
# Testing
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q --cov=ros_to_markdown"
markers = ["integration: marks tests that interact with a real ROS environment"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
################################################################################


################################################################################
# Type Checking
[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
exclude = [
    'docker/',
    'scripts/',
    'tests/',
    '.*/conftest.py', # Exclude all conftest.py files from type checking
]

[[tool.mypy.overrides]]
module = ["rclpy.*", "rospy.*", "rostopic.*", "pydantic.*", "structlog.*", "rosgraph.*"]
ignore_missing_imports = true
################################################################################


################################################################################
# Scripts
[project.scripts]
ros-to-markdown = "ros_to_markdown.cli:main"
################################################################################
